feat: make publishers named "_" skip the same-repo check
Publishers named "_"" are special and get to skip the additional checks,
like the publiccode.yml's url needing to be on the same repo.

They can be used to add repositories and organizations, under the
crawler's admins control, that describe arbitrary repos
(eg. metarepos for other entities)

This is an hack while we come up with a more elegant solution, after
the crawler's refactor.