feat: create Software even if publiccode.yml is invalid
Previously we only created a new Software entity only when the
publiccode.yml is valid, so in case of a new repo with an invalid
publiccode.yml nothing was added.

This means the new software didn't exist from the API's point of view,
and because of that publiccode-issueopener can't notify the maintainers about
the errors, even if new repos are arguably the ones where error reporting and
a little hand holding are more valuable.

So add the software even if publiccode.yml is invalid, setting active to
false.

We will re-activate software whenever publiccode.yml is valid *only*
if it was previously added automatically for the first time as inactive.